Organize cache-related codes into utils/cache.ts
