--- graphql-server/src/utils/oncotree.ts
@@ -1,87 +1,0 @@
-import fetch from "node-fetch";
-import NodeCache from "node-cache";
-import { props } from "./constants";
-import { neo4jDriver } from "./servers";
-
-export const ONCOTREE_CACHE_KEY = "oncotree";
-
-/**
- * Source: https://oncotree.mskcc.org/#/home?tab=api
- */
-export type OncotreeApiTumorType = {
-  children: Record<string, unknown>;
-  code: string;
-  color: string;
-  externalReferences: Record<string, unknown>;
-  history: string[];
-  level: number;
-  mainType: string;
-  name: string;
-  parent: string;
-  precursors: string[];
-  revocations: string[];
-  tissue: string;
-};
-
-export type OncotreeCache = Record<string, { name: string; mainType: string }>;
-
-async function fetchOncotreeApiData() {
-  try {
-    const response = await fetch(props.oncotree_api, {
-      headers: {
-        Accept: "application/json",
-      },
-    });
-    if (!response.ok) {
-      throw new Error(
-        `Failed to fetch the Oncotree API: ${response.statusText}`
-      );
-    }
-    return (await response.json()) as Promise<OncotreeApiTumorType[]>;
-  } catch (error) {
-    if (error instanceof Error) {
-      console.error(`Error: ${error.message}`);
-    } else {
-      console.error("Unknown error occurred while fetching Oncotree data");
-    }
-    return null;
-  }
-}
-
-export async function updateOncotreeCache(inMemoryCache: NodeCache) {
-  const oncotreeApiData = await fetchOncotreeApiData();
-  if (!oncotreeApiData) return;
-  const oncotreeCache = oncotreeApiData.reduce((acc, tumor) => {
-    acc[tumor.code] = { name: tumor.name, mainType: tumor.mainType };
-    return acc;
-  }, {} as OncotreeCache);
-  // Add to cache oncotree codes found in Neo4j but not in Oncotree API's response
-  // (These codes are likely old codes that have been renamed in the Oncotree database)
-  const oncotreeCodesInNeo4j = await getOncotreeCodesFromNeo4j();
-  oncotreeCodesInNeo4j?.forEach((code) => {
-    if (!oncotreeCache.hasOwnProperty(code)) {
-      oncotreeCache[code] = { name: "N/A", mainType: "N/A" };
-    }
-  });
-  inMemoryCache.set(ONCOTREE_CACHE_KEY, oncotreeCache, 84600); // 1 day
-}
-
-async function getOncotreeCodesFromNeo4j() {
-  const session = neo4jDriver.session();
-  try {
-    const result = await session.writeTransaction((tx) =>
-      tx.run(`
-        MATCH (sm:SampleMetadata) RETURN DISTINCT sm.oncotreeCode AS oncotreeCode
-      `)
-    );
-    return result.records
-      .map((record) => record.get("oncotreeCode"))
-      .filter(Boolean) as string[];
-  } catch (error) {
-    if (error instanceof Error) {
-      console.error(error.message);
-    }
-  } finally {
-    await session.close();
-  }
-}
