+++ graphql-server/src/utils/servers.ts
@@ -14,18 +14,8 @@ import {
 import { updateActiveUserSessions } from "./session";
 import { corsOptions } from "./constants";
 import NodeCache from "node-cache";
-import {
-  OncotreeCache,
-  updateOncotreeCache,
-  ONCOTREE_CACHE_KEY,
-} from "./oncotree";
+import { initializeInMemoryCache } from "./cache";
 import neo4j from "neo4j-driver";
-import {
-  buildSamplesQueryBody,
-  buildSamplesQueryFull,
-  queryDashboardSamples,
-} from "../schemas/queries/samples";
-import { DashboardRecordSort, DashboardSample } from "../generated/graphql";
 
 export function initializeHttpsServer(app: Express) {
   return https.createServer(
@@ -55,6 +45,7 @@ export async function initializeApolloServer(
   httpsServer: https.Server,
   app: Express
 ) {
+  console.info("Building, generating, and merging schemas...");
   const { neo4jDbSchema, ogm } = await buildNeo4jDbSchema();
   const customSchema = await buildCustomSchema(ogm);
   const oracleDbSchema = await buildOracleDbSchema();
@@ -62,7 +53,7 @@ export async function initializeApolloServer(
     schemas: [neo4jDbSchema, oracleDbSchema, customSchema],
   });
 
-  const inMemoryCache = await setUpTheInMemoryCache();
+  const inMemoryCache = await initializeInMemoryCache();
 
   const apolloServer = new ApolloServer<ApolloServerContext>({
     schema: mergedSchema,
@@ -88,58 +79,4 @@ export async function initializeApolloServer(
   await apolloServer.start();
   apolloServer.applyMiddleware({ app, cors: corsOptions });
   return apolloServer;
-}
-
-export const SAMPLES_CACHE_KEY = "samples";
-
-async function setUpTheInMemoryCache() {
-  const inMemoryCache = new NodeCache();
-  await updateOncotreeCache(inMemoryCache);
-  await updateSamplesCache(inMemoryCache);
-  // Functions to run when items in cache expire
-  // (node-cache checks for expired items and runs this event listener every 10m by default)
-  inMemoryCache.on("expired", (key) => {
-    if (key === ONCOTREE_CACHE_KEY) {
-      console.info("Refreshing the Oncotree cache...");
-      updateOncotreeCache(inMemoryCache);
-    }
-    if (key === SAMPLES_CACHE_KEY) {
-      console.info("Refreshing the samples cache...");
-      updateSamplesCache(inMemoryCache);
-    }
-  });
-  return inMemoryCache;
-}
-
-// TODO: find an appropriate place to put the code below
-export type SamplesCache = Record<string, DashboardSample[]>;
-
-async function updateSamplesCache(inMemoryCache: NodeCache) {
-  const samplesCache = {} as SamplesCache;
-  const queryBody = buildSamplesQueryBody({
-    searchVals: [],
-    context: undefined,
-    filters: undefined,
-    addlOncotreeCodes: [],
-  });
-  const numOfPagesToCache = 5;
-  const agGridCacheBlockSize = 100; // keep this consistent with helpers.tsx's CACHE_BLOCK_SIZE
-  for (let i = 0; i < numOfPagesToCache; i++) {
-    const offset = i * agGridCacheBlockSize; // [0, 100, 200, 300, 400]
-    const samplesCypherQuery = await buildSamplesQueryFull({
-      queryBody,
-      // Keep this consistent with SamplesList.tsx's DEFAULT_SORT
-      sort: { colId: "importDate", sort: "desc" } as DashboardRecordSort,
-      limit: agGridCacheBlockSize,
-      offset: offset,
-    });
-    const queryResult = await queryDashboardSamples({
-      samplesCypherQuery,
-      oncotreeCache: inMemoryCache.get(ONCOTREE_CACHE_KEY),
-    });
-    if (queryResult) {
-      samplesCache[samplesCypherQuery] = queryResult;
-    }
-  }
-  inMemoryCache.set(SAMPLES_CACHE_KEY, samplesCache, 360); // expires every 1 hour
 }
